External CDDL files and include for #244
This change externalizes the two normative definitions for transparent
and signed statements, respectively, and adds them to the final draft
with the kramdown `includes` directive. This change allows developers
to use a git submodule or programatically use the git version of the
spec to have utilities for code generation and other capabilities use
the definitions without more cumbersome extraction routines for the
source Markdown or RFC XML representations of the spec.