Update the definition of Receipts
* Emphasise the role of Receipts in the abstract definition of
  Transparency
* Specifically acknowlege that signed proofs for other purposes are not
  excluded.
* Emphasise that the specific cryptographic proof is a transparency
  service implementation detail.

The effort to both reduce the definition and also foreshadow the role of
receipts and non-exclusivity of the proof types, seems to be about a net 0 line count
change.

This interpretation seems at odds with blocking on COSE-Receipts, but to
me consistent with the definition of transparency in the document.