retain the link to the macro concept of append only log
The introduction of "Statement Series" is intended to make it clear that
Auditors can verify an entire series of statements "in whole", thus
verifying "append only", without regard to how the transparent
statements were made available for checking.

Note: I had not previosly spotted the requirement in Verifiable Data
Structure that the VDS beable to prove the append only property if used
for a scitt TS. Because of this, I would support droping Append-only in
all other places *except* that one. And could live without "Statement
Series". That said, I believe for systems concerned with provenance, the
story told by an ordered collection of statements, this is important and
useful